added form input test
