+++ src/client/components/FormInput/FormInput.test.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import { render, screen, fireEvent } from '@testing-library/react';
+import { MemoryRouter } from 'react-router-dom';
+import FormInput from './FormInput';
+import LogIn from '../LogIn/LogIn';
+
+describe('FormInput', () => {
+  it('updates input values on change', () => {
+    render(<LogIn />, { wrapper: MemoryRouter });
+    const emailInput = screen.getByLabelText('Email');
+    const passwordInput = screen.getByLabelText('Password');
+
+    fireEvent.change(emailInput, { target: { value: 'test@example.com' } });
+    fireEvent.change(passwordInput, { target: { value: 'password123' } });
+
+    expect(emailInput.value).toBe('test@example.com');
+    expect(passwordInput.value).toBe('password123');
+  });
+});
