+++ src/client/components/HomePage/HomePage.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState, createContext } from 'react';
 
-const HomePage = () => {
+const HomePage: React.FC = () => {
   return <h1>Hi! This is homepage</h1>;
 };
