+++ src/client/components/NavBar/NavBar.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { Link } from 'react-router-dom';
 
-const NavBar = () => {
+const NavBar: React.FC = () => {
   return (
     <Link to="login">
       <button>Log In</button>
